+++ client/src/queries/order/approvalOrder.queries.ts
@@ -1,0 +1,12 @@
+import { gql } from '@apollo/client';
+
+const APPROVAL_ORDER = gql`
+  mutation ApprovalOrder($orderId: String!) {
+    approvalOrder(orderId: $orderId) {
+      result
+      error
+    }
+  }
+`;
+
+export default APPROVAL_ORDER;
