+++ client/src/queries/order/getCompletedOrder.queries.ts
@@ -1,0 +1,27 @@
+import { gql } from '@apollo/client';
+
+const GET_COMPLETED_ORDER = gql`
+  query GetCompletedOrders {
+    getCompletedOrders {
+      result
+      completedOrders {
+        _id
+        startingPoint {
+          coordinates
+          address
+        }
+        destination {
+          coordinates
+          address
+        }
+        amount
+        startedAt
+        completedAt
+        driver
+      }
+      error
+    }
+  }
+`;
+
+export default GET_COMPLETED_ORDER;
