+++ client/src/queries/order/getOrderCarInfo.queries.ts
@@ -1,0 +1,16 @@
+import { gql } from '@apollo/client';
+
+const GET_ORDER_CAR_INFO = gql`
+  query GetOrderCarInfo($orderId: String!) {
+    getOrderCarInfo(orderId: $orderId) {
+      result
+      error
+      carInfo {
+        carNumber
+        carType
+      }
+    }
+  }
+`;
+
+export default GET_ORDER_CAR_INFO;
