+++ client/src/queries/order/index.ts
@@ -1,0 +1,13 @@
+export { default as APPROVAL_ORDER } from './approvalOrder.queries';
+export { default as CANCEL_ORDER } from './cancelOrder.queries';
+export { default as COMPLETE_ORDER } from './completeOrder.queries';
+export { default as CREATE_ORDER } from './createOrder.queries';
+export { default as GET_COMPLETED_ORDER } from './getCompletedOrder.queries';
+export { default as GET_ORDER } from './getOrder.queries';
+export { default as GET_ORDER_BY_ID } from './getOrderById.queries';
+export { default as GET_ORDER_CAR_INFO } from './getOrderCarInfo.queries';
+export { default as GET_UNASSIGNED_ORDERS } from './getUnassignedOrders.queries';
+export { default as START_DRIVING } from './startDriving.queries';
+export { default as SUB_NEW_ORDER } from './subNewOrder.queries';
+export { default as SUB_ORDER_CALL_STATUS } from './subOrderCallStatus.queries';
+export { default as UPDATE_ORDER_LIST } from './updateOrderList.queries';
