+++ client/src/queries/order/startDriving.queries.ts
@@ -1,0 +1,12 @@
+import { gql } from '@apollo/client';
+
+const START_DRIVING = gql`
+  mutation StartDriving($orderId: String!) {
+    startDriving(orderId: $orderId) {
+      result
+      error
+    }
+  }
+`;
+
+export default START_DRIVING;
