+++ client/src/routes/Driver/GoToDestination/GoToDestination.tsx
@@ -5,7 +5,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import MapFrame from '@/components/MapFrame';
 import Modal from '@components/Modal';
 import { UPDATE_DRIVER_LOCATION } from '@queries/user.queries';
-import { COMPLETE_ORDER, GET_ORDER_BY_ID } from '@queries/order.queries';
+import { COMPLETE_ORDER, GET_ORDER_BY_ID } from '@/queries/order';
 import {
   UpdateDriverLocation,
   CompleteOrder,
